Data.Configuration

Type Alias

type ReplacementActionName = Text

type ReplacementSearchString = Text

type TemplateName = Text

Data Types

Text Replacement Action

data TextReplacementAction

A text replacement action to perform on the contents of a file, as read from a configuration file.

Constructors

TextReplacementAction

Fields

_replacementActionName :: ReplacementActionName

The unique name for the replacement action

_replacementActionSearchString :: ReplacementSearchString

The String to find and replace with a file

_replacementText :: Text

The value to substitute in place of the found search string. There are some special values that can be used based on the current directory:

_replacementText Value	Substituted Value

%%THECOURSE%%	The course code found in the current directory.

%%THEASSESSNO%%	The numeric value of the current assessment folder.

%%THEAUTHOR%% 	The author as specified in the used template. If no author is given, no replacement will be made.

Any other values will be substituted as is.

replacementActionName :: Lens' TextReplacementAction ReplacementActionName

replacementActionSearchString :: Lens' TextReplacementAction ReplacementSearchString

replacementText :: Lens' TextReplacementAction Text

Template Configuration

data AuthorInfo

Information about authors for TemplateConfigurations.

Constructors

SingularAuthor Text
A single author

MultipleAuthors (NonEmpty Text)
A non-empty collection of authors

data TemplateConfiguration

A template to specify a directory of files to copy to a new location.

Constructors

TemplateConfiguration

Fields

_authorInfo :: Maybe AuthorInfo

The author of the current template

_templateDirectoryPath :: Path Abs Dir

The absolute path to the directory of files consisting of the template

_relativeOutputPath :: Path Rel Dir

The path relative to the assessment root (or current directory when using --here) where the contents of the template folder will be copied into. An empty path will copy directly into the root directory, without a new folder.

_fileContentReplacementMap :: Map (Path Rel File) (Seq Text)

A mapping of files relative to the template directory with TextReplacementActions to perform. Keys are the relative file paths and values are `Seq Text` of names to TextReplacementActions.

authorInfo :: Lens' TemplateConfiguration (Maybe AuthorInfo)

templateDirectoryPath :: Lens' TemplateConfiguration (Path Abs Dir)

relativeOutputPath :: Lens' TemplateConfiguration (Path Rel Dir)

fileContentReplacementMap :: Lens' TemplateConfiguration (Map (Path Rel File) (Seq Text))

Configuration

data Configuration

A template configuration used to copy a directory when creating a new assessment.

Constructors

Configuration

Fields

_contentReplacementActions :: Seq TextReplacementAction

A collection of replacement actions to perform on specified file contents

_templateConfigurations :: Map TemplateName TemplateConfiguration

A mapping of template configuration names to configurations

contentReplacementActions :: Lens' Configuration (Seq TextReplacementAction)

templateConfigurations :: Lens' Configuration (Map TemplateName TemplateConfiguration)

Other datatypes

data ConfigurationPath

Constructors

NewConfiguration (Path Abs File)

ExistingConfiguration (Path Abs File)

configurationPath :: ConfigurationPath -> Path Abs File

isNewConfiguration :: ConfigurationPath -> Bool