Data.Options

Description

Module contains functionality for parsing command line arguments

Types

These types encapsulates the provided command line arguments.

data ProgramCommand

Datatype for which command the program will execute.

Constructors

CreateAssessment
Create a new assessment directory

Fields

AssessmentCopyToLocation

Whether or not to copy an assessment template here, or into a new assessment directory (via *--here* flag)

GradingCLI
Start the CLI to add, edit, or use hypothetical grades for various features

data AssessmentCopyToLocation

Data type for --here flag. See parseFlagAssessmentCopyToLocation.

Constructors

AssessmentsRootDir

CurrentWorkingDirectory

newtype AppArgs

Data type containing data for parsed command line arguments.

Constructors

AppArgs

Fields

_programCommand :: ProgramCommand

programCommand :: Iso' AppArgs ProgramCommand

data ProgramOptions

Options used at runtime containing parsed AppArgs

Constructors

ProgramOptions

Fields

_appArgs :: AppArgs

_currentWorkingDir :: Path Abs Dir

appArgs :: Lens' ProgramOptions AppArgs

currentWorkingDir :: Lens' ProgramOptions (Path Abs Dir)

Parser functions

argParser :: Parser AppArgs

Application argument parser used to parse AppArgs.

parseProgramCommand :: Parser ProgramCommand

Parses a ProgramCommand.

parseCommandGradingCLI :: Mod CommandFields ProgramCommand

Parses grades command for GradingCLI.

parseFlagAssessmentCopyToLocation :: Parser AssessmentCopyToLocation

Parses --here flag for AssessmentCopyToLocation. Using the flag results in Here. Otherwise, NewCourseAssessment.

parseCommandNewAssessment :: Mod CommandFields ProgramCommand

Parses assess command for CreateAssessment.

Option Parsing

getProgramOptions :: IO ProgramOptions

Used to create a ProgramOptions by parsing AppArgs from the commandline.