Printing

Documentation

type Message = String

type URI = String

type SGRStack = StateT (Seq SGR) IO

startSGRScope :: [SGR] -> SGRStack a -> IO a

Start a new SGR scope with the provided initial SGR paramaters.

nestedSGRScope :: [SGR] -> SGRStack a -> SGRStack a

Start an inner SGR scope, appending the provided options to the stack. These options are set in order, after previously provided options.

sgrHyperlink :: URI -> Message -> SGRStack ()

sgrHyperlinkFile :: URI -> SGRStack ()

sgrGray :: [SGR]

sgrYellow :: [SGR]

sgrBlue :: [SGR]

sgrRed :: [SGR]

sgrPutStr :: Message -> SGRStack ()

sgrPutStrLn :: Message -> SGRStack ()

debug :: String -> IO ()